requirements.txt
# Core dependencies
python-dotenv>=0.20.0
requests>=2.28.1
websocket-client>=1.3.2
numpy>=1.23.0

# CLI interface
rich>=12.5.1
prompt-toolkit>=3.0.30

# Web interface
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.7
starlette>=0.26.1
gunicorn>=20.1.0

# Voice recognition (Core components)
SpeechRecognition>=3.8.1
pyttsx3>=2.90

# Optional dependencies
# To install optional components: pip install -r requirements-optional.txt

# Development and testing
pytest>=7.1.2
pytest-cov>=3.0.0
black>=22.6.0
isort>=5.10.1

# For logging and monitoring
pytz==2023.3 